using core v2 branch and remove v1
