+++ pyproject.toml
@@ -14,7 +14,7 @@ license = "BSD-3-Clause"
 [tool.poetry.dependencies]
 python = "^3.9"
 typing-extensions = "^4.4.0"
-theine-core = {path = "../theine-core", develop = false}
+theine-core = {git = "git@github.com:Yiling-J/theine-core.git", branch = "v2"}
 mypy-extensions = "^1.0.0"
 
 [tool.poetry.group.dev.dependencies]
