+++ theine/adapters/django.py
@@ -1,11 +1,11 @@
 from datetime import timedelta
 from threading import Lock
-from typing import Any, Callable, Dict, Generic, Optional, Type, TypeVar, Union, cast
+from typing import Any, Callable, Dict, Optional, Union, cast
 
 from django.core.cache.backends.base import BaseCache, DEFAULT_TIMEOUT
 
-from theine.v1 import Cache as Theine
-from theine.v1.theine import sentinel
+from theine import Cache as Theine
+from theine.theine import sentinel
 
 KEY_TYPE = Union[str, Callable[..., str]]
 VALUE_TYPE = Any
@@ -15,9 +15,7 @@ VERSION_TYPE = Optional[int]
 class Cache(BaseCache):
     def __init__(self, name: str, params: Dict[str, Any]):
         super().__init__(params)
-        options = params.get("OPTIONS", {})
-        policy = options.get("POLICY", "tlfu")
-        self.cache = Theine(policy, self._max_entries)
+        self.cache = Theine[Any, Any](self._max_entries)
 
     def _timeout_seconds(
         self, timeout: "Optional[Union[float, DEFAULT_TIMEOUT]]"
@@ -55,7 +53,10 @@ class Cache(BaseCache):
         version: VERSION_TYPE = None,
     ) -> Optional[VALUE_TYPE]:
         key = self.make_key(key, version)
-        return self.cache.get(key, default)
+        v, ok = self.cache.get(key)
+        if not ok:
+            return default
+        return v
 
     def set(
         self,
