--- theine/v1/__init__.py
@@ -1,5 +1,0 @@
-from theine_core.v1 import BloomFilter
-
-from .theine import Cache, Memoize
-
-__all__ = ("Cache", "BloomFilter", "Memoize")
