--- theine/v2/__init__.py
@@ -1,5 +1,0 @@
-from theine_core.v2 import BloomFilter
-
-from .theine import Cache, Memoize
-
-__all__ = ("Cache", "BloomFilter", "Memoize")
