--- theine/v2/exceptions.py
@@ -1,2 +1,0 @@
-class InvalidTTL(Exception):
-    pass
